+++ src/game/mod.rs
@@ -1,15 +1,27 @@
 //! Game mechanics and content.
+//!
+//! The basic movement code shipped with the template is taken from the corresponding
+//! [Bevy example](https://github.com/janhohenheim/bevy/blob/fixed-time-movement/examples/movement/physics_in_fixed_timestep.rs).
+//! See that link for an in-depth explanation of the code and the motivation behind it.
 
 use bevy::prelude::*;
-pub(crate) use spawn_level::SpawnLevel;
+pub(crate) use level::SpawnLevel;
 
+mod level;
 mod movement;
-mod spawn_level;
+mod physics;
+mod render;
 
 pub(super) fn plugin(app: &mut App) {
     app.configure_sets(Update, GameSystem::Movement);
 
-    app.add_plugins((spawn_level::plugin, movement::plugin));
+    app.add_plugins((
+        movement::plugin,
+        physics::plugin,
+        render::plugin,
+        level::plugin,
+        movement::plugin,
+    ));
 }
 
 #[derive(Debug, SystemSet, Clone, Copy, Eq, PartialEq, Hash)]
